Add dependency on `libxmljs`
After experimenting with lots of different approaches, I think working with proper `libxml` bindings is probably best (esp. since I have good familiarity with `libxml`).  It remains possible that another solution will have to be found (that doesn't depend on compiling C libraries, perhaps), but in the interests of getting on with it, I'm choosing this now.