Allow node to import ES6 modules
This commit adds the `esm` module and requires it on all calls to the gatsby-cli.  This allows using ES6 modules and import syntax in (most specifically) `gatsby-node.js`, so that modules like the JsonMlDoc and TeiDoc can be imported this way.  This is important because I want to import them this way in the client-side code, to ensure that code-splitting will work properly.